third_party/blink/web_tests/fast/peerconnection/RTCSessionDescription-expected.txt

Tests RTCSessionDescription.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS sessionDescription = new RTCSessionDescription(initializer); did not throw exception.
PASS sessionDescription.type is "offer"
PASS sessionDescription.sdp is "foobar"
PASS initializer = JSON.parse(JSON.stringify(sessionDescription)); did not throw exception.
PASS sessionDescription = new RTCSessionDescription(initializer); did not throw exception.
PASS sessionDescription.type is "offer"
PASS sessionDescription.sdp is "foobar"
PASS new RTCSessionDescription(); did not throw exception.
PASS new RTCSessionDescription({}); did not throw exception.
PASS new RTCSessionDescription(5); threw exception TypeError: Failed to construct 'RTCSessionDescription': The provided value is not of type 'RTCSessionDescriptionInit'..
PASS new RTCSessionDescription('foobar'); threw exception TypeError: Failed to construct 'RTCSessionDescription': The provided value is not of type 'RTCSessionDescriptionInit'..
PASS new RTCSessionDescription({type:'foobar', sdp:'x'}); threw exception TypeError: Failed to construct 'RTCSessionDescription': Failed to read the 'type' property from 'RTCSessionDescriptionInit': The provided value 'foobar' is not a valid enum value of type RTCSdpType..
PASS new RTCSessionDescription({type:'offer', sdp:''}); did not throw exception.
PASS new RTCSessionDescription({type:'offer', sdp:'x'}); did not throw exception.
PASS new RTCSessionDescription({type:'answer', sdp:'x'}); did not throw exception.
PASS new RTCSessionDescription({type:'pranswer', sdp:'x'}); did not throw exception.
PASS successfullyParsed is true

TEST COMPLETE
